changes on the parameters type

+++ app_backend/app/models/vessel.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, Integer, String,Float
 from ..database import Base
 from sqlalchemy.orm import relationship
 
@@ -10,6 +10,6 @@ class Vessel(Base):
     name = Column(String, nullable=False)
     registration_number = Column(String, unique=True, nullable=False)
     capacity = Column(Integer, nullable=True)
-    tonnage = Column(Integer, nullable=True)
+    tonnage = Column(Float, nullable=True)
 
     documents = relationship("Document", back_populates="vessel", cascade="all, delete")
